requirements.txt
Click==7.0
dash==1.9.0
dash-bootstrap-components==0.9.1
dash-core-components==1.8.0
dash-html-components==1.0.2
dash-renderer==1.2.4
dash-table==4.6.0
Flask==1.1.1
Flask-Compress==1.4.0
future==0.18.2
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
numpy==1.18.1
pkg-resources==0.0.0
plotly==4.5.1
PyYAML==5.3
retrying==1.3.3
RPi.GPIO==0.7.0
six==1.14.0
spidev==3.4
tqdm==4.43.0
uWSGI==2.0.18
Werkzeug==1.0.0
pandas
